Create an improved function search macro, and use it to look for res_query.
Instead of using AC_SEARCH_LIBS or AC_CHECK_LIB, create a new macro to
search for function in different libraries; this is needed since the
function name might just be a (preprocessor) macro for another symbol.

The new macro is released under GPL2 (or later) + exception, and
allows to provide a custom source code to use to test for the
function.